[Refactor] #40 - ItemDecoration 빼내기
